# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: tokrouter API
  description: |
    LLM API Gateway and Router.

    tokrouter provides a unified interface for multiple LLM providers with:
    - Protocol conversion (OpenAI ↔ Anthropic)
    - Automatic fallback and retry
    - Circuit breaker for endpoint health
    - Cost tracking and usage statistics

    Configure providers in config.yaml and start with `tokrouter serve`.
  version: 0.7.0
  contact:
    name: tokrouter
    url: https://github.com/tokzone/tokrouter
servers:
  - url: http://127.0.0.1:8765
    description: Local development server
paths:
...
  /v1/messages:
    post:
      summary: Messages (Anthropic format)
      description: |
        Send a message request using Anthropic's API format.
        Requests are forwarded to configured providers with automatic protocol conversion.
      operationId: messages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessagesRequest'
            examples:
              basic:
                summary: Basic request
                value:
                  model: claude-3-5-sonnet-20241022
                  max_tokens: 1024
                  messages:
                    - role: user
                      content: Hello!
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
            text/event-stream:
              schema:
                type: string
                description: SSE stream of message chunks
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Upstream error or no available endpoints
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    MessagesRequest:
      type: object
      required:
        - model
        - max_tokens
        - messages
      properties:
        model:
          type: string
          description: Model name
          example: claude-3-5-sonnet-20241022
        max_tokens:
          type: integer
          description: Maximum tokens in response
        messages:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
                enum: [user, assistant]
              content:
                type: string
        system:
          type: string
          description: System prompt
        stream:
          type: boolean
          default: false
    MessagesResponse:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          example: message
        role:
          type: string
          example: assistant
        content:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              text:
                type: string
        model:
          type: string
        stop_reason:
          type: string
        usage:
          $ref: '#/components/schemas/Usage'
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
          description: Input tokens used
        completion_tokens:
          type: integer
          description: Output tokens generated
        total_tokens:
          type: integer
          description: Total tokens
    Error:
      type: object
      properties:
        type:
          type: string
          example: error
        code:
          type: string
          description: Error code
          enum:
            - INVALID_REQUEST
            - UPSTREAM_FAILED
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
          description: Error message
...